# imadoko_back/src/main/java/com/c/imadoko/imadoko_back/domain/Team.java
package com.c.imadoko.imadoko_back.domain;

import jakarta.persistence.*;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.ArrayList;
import java.util.List;

@Entity
@Data
@NoArgsConstructor
public class Team {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @NotBlank(message = "チーム名は必須です")
    @Size(max = 50, message = "チーム名は50文字以内で入力してください")
    @Column(unique = true, nullable = false)
    private String teamName;

    @Size(max = 14, message = "選手は最大14人まで登録できます")
    @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
    @JoinColumn(name = "team_id")
    private List<Player> players = new ArrayList<>();

    public Team(String teamName) {
        this.teamName = teamName;
    }
}